notifikasi dan export exel
doit

+++ database/migrations/2025_09_11_065035_create_notifikasi_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('notifikasi')) {
+            return;
+        }
+
         Schema::create('notifikasi', function (Blueprint $table) {
             $table->id();
             $table->foreignId('keluhan_id')->nullable()->constrained('keluhan')->onDelete('cascade');
